server/backend/requirements.txt
# AI/ML Core Dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# LLM and Embeddings
google-generativeai==0.3.0
openai==1.3.0
transformers==4.35.0
sentence-transformers==2.2.2

# Vector Database
chromadb==0.4.15
pinecone-client==2.2.4

# NLP Processing
spacy==3.7.2
nltk==3.8.1
googletrans==4.0.0
langdetect==1.0.9
indic-transliteration==2.3.43

# Multimodal Processing
Pillow==10.1.0
opencv-python==4.8.1
librosa==0.10.1
speech_recognition==3.10.0
pydub==0.25.1

# Data Processing
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
faiss-cpu==1.7.4

# Web Framework & Security
fastapi-limiter==0.1.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Database & Caching
motor==3.3.2  # Async MongoDB
redis==5.0.1
pymongo==4.6.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1
httpx==0.25.2